requirements.txt
aiohttp==2.2.3
aiomysql==0.0.9
amqp==2.1.4
arrow==0.10.0
async-timeout==1.2.1
asyncio==3.4.3
billiard==3.5.0.2
celery==4.0.2
chardet==3.0.4
Django==1.11.3
django-cors-headers==2.1.0
django-debug-toolbar==1.8
django-filter==1.0.4
djangorestframework==3.6.3
flake8==3.3.0
importmagic==0.1.7
jedi==0.10.2
kombu==4.0.2
mccabe==0.6.1
multidict==3.1.0
mysqlclient==1.3.10
pycodestyle==2.3.1
pyflakes==1.5.0
PyMySQL==0.7.11
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
sqlparse==0.2.3
vine==1.1.3
yapf==0.16.2
yarl==0.11.0
